Notebook model, including utility for loading from, and saving to, YAML.
Introduces Notebook, our temporary name for a document model.  It holds the location you loaded the binary from and also the annotations.
You can load from, and save to, YAML.  There is no UI for it yet though, and there is no point having any until there is UI for annotations.
The Annotation API has been fleshed out a bit more.
New factory-like API for interpretor providers... needs some work still.  Also needs a lot more interpretors, and when we try to hook it into the GUI we will also need some way of creating a menu of all annotations.
Required a bunch of customisations to JvYAML to get it to work (all through their public API of course.)