Initial implementation of annotation rendering.
Annotations are now rendered in a red border and faded red background, to be customisable later.
They also know their length.  I was originally avoiding this, thinking that the interpreters should know the length of their values, but it turns out if you do this, you always need a reference to the binary to determine the length, and the annotation collection should be able to exist without a reference to a Binary.