Support adding a group around existing annotations.
For this one I extracted out the utility for searching over a set of annotations and am testing it in what should be an exhaustive set of cases.  The results returned are vaguely reminiscent of range maths, but with some cases removed (e.g. if the start matches and the end is higher, this is treated the same as the start being lower and the end being higher.)