Issue #6: Fixes the weird issue on exit where you get asked to save a notebook you already closed
It looks like Java doesn't remove a frame from the list you get from
Frame.getFrames() immediately after you dispose() it.  The workaround is
to check if the notebook was open, but interestingly if you closed the
notebook by closing the window instead of using Close Notebook from the
menu, it wouldn't have actually closed the notebook.  By fixing both, it
works as you would expect.